+++ console2/src/components/organisms/StartRepositoryPopup/index.tsx
@@ -25,21 +25,22 @@ import { Button } from 'semantic-ui-react';
 
 import { ConcordId, ConcordKey, RequestError } from '../../../api/common';
 import { StartProcessResponse } from '../../../api/org/process';
-import { actions, State } from '../../../state/data/processes';
-import { SingleOperationPopup } from '../../molecules';
+import { actions, State as ProcessState } from '../../../state/data/processes';
+import { SingleOperationPopup, GitHubLink } from '../../molecules';
 
-import './styles.css';
+import { Table, Keys, Key, Values, Value, Input } from './styles';
 
 interface ExternalProps {
     orgName: ConcordKey;
     projectName: ConcordKey;
     repoName: ConcordKey;
+    repoURL: string;
     trigger: (onClick: () => void) => React.ReactNode;
 }
 
 interface DispatchProps {
     reset: () => void;
-    onConfirm: () => void;
+    onConfirm: (entryPoint: string) => void;
     openProcessPage: (instanceId: ConcordId) => void;
 }
 
@@ -52,7 +53,22 @@ interface StateProps {
 
 type Props = DispatchProps & ExternalProps & StateProps;
 
-class StartRepositoryPopup extends React.Component<Props> {
+interface State {
+    entryPoint: string;
+}
+
+class StartRepositoryPopup extends React.Component<Props, State> {
+    constructor(props: Props) {
+        super(props);
+        this.state = {
+            entryPoint: 'default'
+        };
+    }
+
+    updateEntryPoint(e: React.KeyboardEvent<HTMLInputElement>) {
+        this.setState({ entryPoint: e.currentTarget.value });
+    }
+
     render() {
         const {
             repoName,
@@ -60,6 +76,7 @@ class StartRepositoryPopup extends React.Component<Props> {
             starting,
             success,
             response,
+            repoURL,
             error,
             reset,
             onConfirm,
@@ -67,7 +84,7 @@ class StartRepositoryPopup extends React.Component<Props> {
         } = this.props;
 
         const successMsg = response ? (
-            <div className="resultBox">
+            <div>
                 {JSON.stringify(
                     {
                         ok: response.ok,
@@ -83,13 +100,35 @@ class StartRepositoryPopup extends React.Component<Props> {
 
         return (
             <SingleOperationPopup
+                customStyle={{ maxWidth: '800px' }}
                 trigger={trigger}
-                title="Start a new process?"
+                title={`Start flow ${repoName}`}
+                icon="triangle right"
+                iconColor="blue"
                 introMsg={
-                    <p>
-                        Are you sure you want to start a new process using <b>{repoName}</b>{' '}
-                        repository?
-                    </p>
+                    <Table>
+                        <Keys>
+                            <Key>Repository URL</Key>
+                            <Key>Source</Key>
+                            <Key>Path</Key>
+                            <Key>Entry Point</Key>
+                        </Keys>
+
+                        <Values>
+                            <Value>
+                                <GitHubLink url={repoURL} text={repoURL} />
+                            </Value>
+                            <Value>master</Value>
+                            <Value>/</Value>
+                            <Value>
+                                <Input
+                                    type="text"
+                                    placeholder="default"
+                                    onKeyUp={(e: any) => this.updateEntryPoint(e)}
+                                />
+                            </Value>
+                        </Values>
+                    </Table>
                 }
                 running={starting}
                 runningMsg={<p>Starting the process...</p>}
@@ -97,7 +136,7 @@ class StartRepositoryPopup extends React.Component<Props> {
                 successMsg={successMsg}
                 error={error}
                 reset={reset}
-                onConfirm={onConfirm}
+                onConfirm={() => onConfirm(this.state.entryPoint || 'default')}
                 onDoneElements={() => (
                     <Button
                         basic={true}
@@ -105,12 +144,14 @@ class StartRepositoryPopup extends React.Component<Props> {
                         onClick={() => instanceId && openProcessPage(instanceId)}
                     />
                 )}
+                customYes="Start"
+                customNo="Cancel"
             />
         );
     }
 }
 
-const mapStateToProps = ({ processes }: { processes: State }): StateProps => ({
+const mapStateToProps = ({ processes }: { processes: ProcessState }): StateProps => ({
     starting: processes.startProcess.running,
     success: !!processes.startProcess.response && processes.startProcess.response.ok,
     response: processes.startProcess.response,
@@ -122,7 +163,8 @@ const mapDispatchToProps = (
     { orgName, projectName, repoName }: ExternalProps
 ): DispatchProps => ({
     reset: () => dispatch(actions.reset()),
-    onConfirm: () => dispatch(actions.startProcess(orgName, projectName, repoName)),
+    onConfirm: (entryPoint: string) =>
+        dispatch(actions.startProcess(orgName, projectName, repoName, entryPoint)),
     openProcessPage: (instanceId: ConcordId) => dispatch(pushHistory(`/process/${instanceId}`))
 });